TRAN-270: Fix save to external SD card

+++ domain/src/main/kotlin/mega/privacy/android/domain/usecase/transfers/sd/HandleSDCardEventUseCase.kt
@@ -65,8 +65,8 @@ class HandleSDCardEventUseCase @Inject constructor(
                             )
                         }
                     }
-                    if (transfer.isRootTransfer) {
-                        //if it's a root transfer we can remove the sd transfer from the data base
+                    if (transfer.isRootTransfer && transfer.isSDCardDownload()) {
+                        //if it's a root sd card transfer we can remove the sd transfer from the data base
                         deleteSdTransferByTagUseCase(transferEvent.transfer.tag)
                     }
                 }
